Login form AND Logout form complete
When an account is registered, an email will be sent out and the information will be stored in the DB.

Login form checks to see if the email and password stored in the DB is true, if it is login will be redirected to a new page.

Login information can be remembered with the "RememberMe", if this is clicked the information will be stored within the local cookies for one year, I can change this at a later date.

Logout form logs out the user and redirects to another page.