Admin Privileges
Admin Privileges, struggling to implement is-inrole functionality, not sure why I can't get it to work.

For the mean time I have implemented the same functionality within the _Layout class. Checking whether or not the email address is within a specific database and it matches with the httpcontext.users.identity

This might have some bugs I will figure this out tomorrow.