Reset Password implementation
Forgotten password and reset password implemented, if a user forgets their password AND their email address is within the database. An email will be sent to the email address with a reset password hyperlink, here a reset-code will be printed into the text box allowing a user to add a new password. I have checked old passwords and they do not work, so the reset/forgotten password implementation is working.

Not the best way of securing a users account, but it will do for the time being.