Implemented Current User in Ticket System View
When a user makes a ticket, their email address will be displayed on screen. At the moment, I have implemented a text box that autofills with the users email address, this might be changed to username instead but it works for the time being.

Also changed the Ticket database, "CurrentUser" has been implemented. This may need changing at a later stage.